% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cols.R
\name{fake_column}
\alias{fake_column}
\title{Schema graph_tbl with already faked data (if performed)}
\usage{
fake_column(n, col_def, schema, faker_opts)
}
\arguments{
\item{n}{Number of values to simulate}

\item{col_def}{Column definition.}

\item{schema}{Schema object.}

\item{faker_opts}{Column simulation faker_opts.}
}
\description{
Schema graph_tbl with already faked data (if performed)
}
\keyword{internal}
