% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_conf.R
\docType{data}
\name{faker_configuration}
\alias{faker_configuration}
\alias{default_faker_opts}
\alias{set_faker_opts}
\title{Default options for pulling metadata and data simulation}
\format{
An object of class \code{list} of length 27.
}
\usage{
default_faker_opts

set_faker_opts(
  opt_pull_character,
  opt_pull_numeric,
  opt_pull_integer,
  opt_pull_logical,
  opt_pull_date,
  opt_pull_table,
  opt_default_character,
  opt_simul_spec_character,
  opt_simul_restricted_character,
  opt_simul_default_fun_character,
  opt_default_numeric,
  opt_simul_spec_numeric,
  opt_simul_restricted_numeric,
  opt_simul_default_fun_numeric,
  opt_default_integer,
  opt_simul_spec_integer,
  opt_simul_restricted_integer,
  opt_simul_default_fun_integer,
  opt_default_logical,
  opt_simul_spec_logical,
  opt_simul_restricted_logical,
  opt_simul_default_fun_logical,
  opt_default_date,
  opt_simul_spec_date,
  opt_simul_restricted_date,
  opt_simul_default_fun_date,
  opt_default_table,
  global = TRUE
)
}
\arguments{
\item{opt_pull_character, opt_pull_numeric, opt_pull_integer, opt_pull_logical, opt_pull_date, opt_pull_table, opt_default_character, opt_simul_spec_character, opt_simul_restricted_character, opt_simul_default_fun_character, opt_default_numeric, opt_simul_spec_numeric, opt_simul_restricted_numeric, opt_simul_default_fun_numeric, opt_default_integer, opt_simul_spec_integer, opt_simul_restricted_integer, opt_simul_default_fun_integer, opt_default_logical, opt_simul_spec_logical, opt_simul_restricted_logical, opt_simul_default_fun_logical, opt_default_date, opt_simul_spec_date, opt_simul_restricted_date, opt_simul_default_fun_date, opt_default_table}{Parameters defined in default configuration that can be modified by using \code{set_faker_opts} function.
Please make sure each parameter is specified by method designed to it.}

\item{global}{If TRUE, default the configuration will be set up globally
(no need to pass it as a \code{faker_opts} parameter for \link{schema_source} and \link{schema_methods}).}
}
\description{
Generated with the set of configuration functions:
\link{default_simulation_params}, \link{opt_default_table},
\link{special_simulation}, \link{restricted_simulation},
\link{sourcing_metadata}.
}
\details{
\code{set_faker_opts} allows to overwrite selected options,
}
\keyword{datasets}
